MOVED:  email verification methods from AuthService  to the EmailService class
