+++ services/index.js
@@ -1,8 +1,11 @@
 const AuthService = require("./AuthService");
 
-const ContactsService = require("./ContactsService")
+const ContactsService = require("./ContactsService");
+
+const EmailService = require("./EmailService");
 
 module.exports = {
   AuthService,
-  ContactsService
+  ContactsService,
+  EmailService,
 };
